Finished refactor of visitor pattern to prep for strategy pattern
