+++ include/shape_visitor.hpp
@@ -22,9 +22,54 @@ class ShapeVisitor {
          *  The ShapeVisitor base class must know about all of the
          *  types in advance. 
          **/
-        virtual void visit( shape_visitor::Circle const& ) const = 0;
-        virtual void visit( shape_visitor::Square const& ) const = 0;
-        virtual void visit( shape_visitor::Rectangle const& ) const = 0;
+        virtual void visit( Circle const& ) const = 0;
+        virtual void visit( Square const& ) const = 0;
+        virtual void visit( Rectangle const& ) const = 0;
+};
+
+class Shape {
+    /**
+     *  This class is not implemented - it is an interface to accept visitors.
+     */
+    public:
+        virtual ~Shape() = default;
+        virtual void accept( shape_visitor::ShapeVisitor const& sv ) = 0;
+};
+
+class Circle : public Shape
+{
+    public:
+        explicit Circle( double radius ) : radius_( radius ) {}
+        
+        void accept( shape_visitor::ShapeVisitor const& sv ) override { sv.visit(*this); } 
+        double radius() const { return radius_; }
+
+    private:
+        double radius_;
+};
+
+class Square : public Shape {
+    public:
+        explicit Square( double side ) : side_ (side) {}
+
+        void accept( shape_visitor::ShapeVisitor const& sv ) override { sv.visit(*this); }
+        double side() const { return side_; }
+
+    private:
+        double side_;
+};
+
+class Rectangle : public Shape {
+    public:
+        explicit Rectangle(double length, double width) : length_ (length), width_ (width) {}
+        
+        void accept( shape_visitor::ShapeVisitor const& sv ) override { sv.visit(*this); }
+        double length() const { return length_; }
+        double width() const { return width_; }
+
+    private:
+        double length_;
+        double width_;
 };
 
 }
